renamed chat_bot_booking.py to chat_bot_service_provider.py
